photoalbum/requirements.txt
appdirs==1.4.3
argcomplete==1.9.3
asgiref==3.2.3
attrs==19.3.0
black==19.10b0
boto3==1.10.36
botocore==1.13.36
certifi==2019.11.28
cfn-flip==1.2.2
chardet==3.0.4
Click==7.0
dataclasses==0.6
dj-database-url==0.5.0
Django==2.2.8
django-appconf==1.0.3
django-crispy-forms==1.8.1
django-heroku==0.3.1
django-imagekit==4.0.2
django-storages==1.8
django-suit==0.2.28
djangorestframework==3.11.0
djongo==1.2.38
dnspython==1.16.0
docutils==0.15.2
durationpy==0.5
future==0.16.0
gunicorn==20.0.4
hjson==3.0.1
idna==2.8
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.20.0
pathspec==0.6.0
pilkit==2.0
Pillow==6.2.1
placebo==0.9.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
pymongo==3.10.0
python-dateutil==2.6.1
python-decouple==3.3
python-slugify==1.2.4
pytz==2019.3
PyYAML==5.2
regex==2019.12.9
requests==2.22.0
s3transfer==0.2.1
six==1.13.0
sqlparse==0.2.4
toml==0.10.0
tqdm==4.19.1
troposphere==2.5.3
typed-ast==1.4.0
Unidecode==1.1.1
urllib3==1.25.7
Werkzeug==0.16.0
whitenoise==5.0.1
wsgi-request-logger==0.4.6
